Silence compiler

+++ Plugins/Audio/snd_fx.cpp
@@ -95,117 +95,117 @@ void SX_Init(void)
   // flGain = EAX 1.0 volume
   // flDecayTime = EAX 1.0 decay time
   // flDecayHFRatio = EAX 1.0 damping (probably)
-  presets_room[1].flGain = 0.417;
-  presets_room[1].flDecayTime = 0.4;
-  presets_room[1].flDecayHFRatio = 2 / 3;
+  presets_room[1].flGain = 0.417f;
+  presets_room[1].flDecayTime = 0.4f;
+  presets_room[1].flDecayHFRatio = static_cast<float>(2 / 3);
 
-  presets_room[2].flGain = 0.3;
-  presets_room[2].flDecayTime = 1.5;
-  presets_room[2].flDecayHFRatio = 1 / 6;
+  presets_room[2].flGain = 0.3f;
+  presets_room[2].flDecayTime = 1.5f;
+  presets_room[2].flDecayHFRatio = static_cast<float>(1 / 6);
 
-  presets_room[3].flGain = 0.4;
-  presets_room[3].flDecayTime = 1.5;
-  presets_room[3].flDecayHFRatio = 1 / 6;
+  presets_room[3].flGain = 0.4f;
+  presets_room[3].flDecayTime = 1.5f;
+  presets_room[3].flDecayHFRatio = static_cast<float>(1 / 6);
 
-  presets_room[4].flGain = 0.6;
-  presets_room[4].flDecayTime = 1.5;
-  presets_room[4].flDecayHFRatio = 1 / 6;
+  presets_room[4].flGain = 0.6f;
+  presets_room[4].flDecayTime = 1.5f;
+  presets_room[4].flDecayHFRatio = static_cast<float>(1 / 6);
 
-  presets_room[5].flGain = 0.4;
-  presets_room[5].flDecayTime = 2.886;
+  presets_room[5].flGain = 0.4f;
+  presets_room[5].flDecayTime = 2.886f;
   presets_room[5].flDecayHFRatio = 0.25;
 
-  presets_room[6].flGain = 0.6;
-  presets_room[6].flDecayTime = 2.886;
-  presets_room[6].flDecayHFRatio = 0.25;
+  presets_room[6].flGain = 0.6f;
+  presets_room[6].flDecayTime = 2.886f;
+  presets_room[6].flDecayHFRatio = 0.25f;
 
-  presets_room[7].flGain = 0.8;
-  presets_room[7].flDecayTime = 2.886;
-  presets_room[7].flDecayHFRatio = 0.25;
+  presets_room[7].flGain = 0.8f;
+  presets_room[7].flDecayTime = 2.886f;
+  presets_room[7].flDecayHFRatio = 0.25f;
 
-  presets_room[8].flGain = 0.5;
-  presets_room[8].flDecayTime = 2.309;
-  presets_room[8].flDecayHFRatio = 0.888;
+  presets_room[8].flGain = 0.5f;
+  presets_room[8].flDecayTime = 2.309f;
+  presets_room[8].flDecayHFRatio = 0.888f;
 
-  presets_room[9].flGain = 0.65;
-  presets_room[9].flDecayTime = 2.309;
-  presets_room[9].flDecayHFRatio = 0.888;
+  presets_room[9].flGain = 0.65f;
+  presets_room[9].flDecayTime = 2.309f;
+  presets_room[9].flDecayHFRatio = 0.888f;
 
-  presets_room[10].flGain = 0.8;
-  presets_room[10].flDecayTime = 2.309;
-  presets_room[10].flDecayHFRatio = 0.888;
+  presets_room[10].flGain = 0.8f;
+  presets_room[10].flDecayTime = 2.309f;
+  presets_room[10].flDecayHFRatio = 0.888f;
 
-  presets_room[11].flGain = 0.3;
-  presets_room[11].flDecayTime = 2.697;
-  presets_room[11].flDecayHFRatio = 0.638;
+  presets_room[11].flGain = 0.3f;
+  presets_room[11].flDecayTime = 2.697f;
+  presets_room[11].flDecayHFRatio = 0.638f;
 
-  presets_room[12].flGain = 0.5;
-  presets_room[11].flDecayTime = 2.697;
-  presets_room[11].flDecayHFRatio = 0.638;
+  presets_room[12].flGain = 0.5f;
+  presets_room[11].flDecayTime = 2.697f;
+  presets_room[11].flDecayHFRatio = 0.638f;
 
-  presets_room[13].flGain = 0.65;
-  presets_room[11].flDecayTime = 2.697;
-  presets_room[11].flDecayHFRatio = 0.638;
+  presets_room[13].flGain = 0.65f;
+  presets_room[11].flDecayTime = 2.697f;
+  presets_room[11].flDecayHFRatio = 0.638f;
 
-  presets_room[14].flGain = 1;
-  presets_room[14].flDecayTime = 1.5;
-  presets_room[14].flDecayHFRatio = 0;
+  presets_room[14].flGain = 1.0f;
+  presets_room[14].flDecayTime = 1.5f;
+  presets_room[14].flDecayHFRatio = 0.0f;
 
-  presets_room[15].flGain = 1;
-  presets_room[15].flDecayTime = 2.5;
-  presets_room[15].flDecayHFRatio = 0;
+  presets_room[15].flGain = 1.0f;
+  presets_room[15].flDecayTime = 2.5f;
+  presets_room[15].flDecayHFRatio = 0.0f;
 
-  presets_room[16].flGain = 1;
-  presets_room[16].flDecayTime = 3.5;
-  presets_room[16].flDecayHFRatio = 0;
+  presets_room[16].flGain = 1.0f;
+  presets_room[16].flDecayTime = 3.5f;
+  presets_room[16].flDecayHFRatio = 0.0f;
 
-  presets_room[17].flGain = 0.65;
-  presets_room[17].flDecayTime = 1.493;
-  presets_room[17].flDecayHFRatio = 0.5;
+  presets_room[17].flGain = 0.65f;
+  presets_room[17].flDecayTime = 1.493f;
+  presets_room[17].flDecayHFRatio = 0.5f;
 
-  presets_room[18].flGain = 0.85;
-  presets_room[18].flDecayTime = 1.493;
-  presets_room[18].flDecayHFRatio = 0.5;
+  presets_room[18].flGain = 0.85f;
+  presets_room[18].flDecayTime = 1.493f;
+  presets_room[18].flDecayHFRatio = 0.5f;
 
-  presets_room[19].flGain = 1;
-  presets_room[19].flDecayTime = 1.493;
-  presets_room[19].flDecayHFRatio = 0.5;
+  presets_room[19].flGain = 1.0f;
+  presets_room[19].flDecayTime = 1.493f;
+  presets_room[19].flDecayHFRatio = 0.5f;
 
-  presets_room[20].flGain = 0.4;
-  presets_room[20].flDecayTime = 7.284;
-  presets_room[20].flDecayHFRatio = 1 / 3;
+  presets_room[20].flGain = 0.4f;
+  presets_room[20].flDecayTime = 7.284f;
+  presets_room[20].flDecayHFRatio = static_cast<float>(1 / 3);
 
-  presets_room[21].flGain = 0.55;
-  presets_room[21].flDecayTime = 7.284;
-  presets_room[21].flDecayHFRatio = 1 / 3;
+  presets_room[21].flGain = 0.55f;
+  presets_room[21].flDecayTime = 7.284f;
+  presets_room[21].flDecayHFRatio = static_cast<float>(1 / 3);
 
-  presets_room[22].flGain = 0.7;
-  presets_room[22].flDecayTime = 7.284;
-  presets_room[22].flDecayHFRatio = 1 / 3;
+  presets_room[22].flGain = 0.7f;
+  presets_room[22].flDecayTime = 7.284f;
+  presets_room[22].flDecayHFRatio = static_cast<float>(1 / 3);
 
-  presets_room[23].flGain = 0.5;
-  presets_room[23].flDecayTime = 3.961;
-  presets_room[23].flDecayHFRatio = 0.5;
+  presets_room[23].flGain = 0.5f;
+  presets_room[23].flDecayTime = 3.961f;
+  presets_room[23].flDecayHFRatio = 0.5f;
 
-  presets_room[24].flGain = 0.7;
-  presets_room[24].flDecayTime = 3.961;
-  presets_room[24].flDecayHFRatio = 0.5;
+  presets_room[24].flGain = 0.7f;
+  presets_room[24].flDecayTime = 3.961f;
+  presets_room[24].flDecayHFRatio = 0.5f;
 
-  presets_room[25].flGain = 1;
-  presets_room[25].flDecayTime = 3.961;
-  presets_room[25].flDecayHFRatio = 0.5;
+  presets_room[25].flGain = 1.0f;
+  presets_room[25].flDecayTime = 3.961f;
+  presets_room[25].flDecayHFRatio = 0.5f;
 
-  presets_room[26].flGain = 0.2;
-  presets_room[26].flDecayTime = 17.234;
-  presets_room[26].flDecayHFRatio = 2 / 3;
+  presets_room[26].flGain = 0.2f;
+  presets_room[26].flDecayTime = 17.234f;
+  presets_room[26].flDecayHFRatio = static_cast<float>(2 / 3);
 
-  presets_room[27].flGain = 0.3;
-  presets_room[27].flDecayTime = 17.234;
-  presets_room[27].flDecayHFRatio = 2 / 3;
+  presets_room[27].flGain = 0.3f;
+  presets_room[27].flDecayTime = 17.234f;
+  presets_room[27].flDecayHFRatio = static_cast<float>(2 / 3);
 
-  presets_room[28].flGain = 0.4;
-  presets_room[28].flDecayTime = 17.234;
-  presets_room[28].flDecayHFRatio = 2 / 3;
+  presets_room[28].flGain = 0.4f;
+  presets_room[28].flDecayTime = 17.234f;
+  presets_room[28].flDecayHFRatio = static_cast<float>(2 / 3);
 
   // Init each effect
   for (int i = 0; i < CSXROOM; ++i)
